requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
motor>=3.0.0
pymongo>=4.0.0
python-jose[cryptography]>=3.0.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.5
python-dotenv>=1.0.0
openai>=1.0.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
Pillow>=9.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-decouple>=3.0.0
aiofiles>=23.0.0
aiosmtplib>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0